requirements.txt
Flask==3.1.1
requests==2.32.4
mercadopago==2.3.0
pyTelegramBotAPI==4.27.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
Werkzeug==3.1.3
